requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
bcrypt==4.1.2
cryptography==41.0.7
gunicorn==21.2.0
python-dotenv